Migrate config loading to internal/cli/config (#360)
Move FileConfig, FileConfigPath, LoadFileConfig, and the provider-agnostic
GenericBaseURLEnvVar name derivation into internal/cli/config, exported.
cmd/pigo now references the new package; applyFileConfig stays in cmd/pigo
(consumes config.FileConfig) pending the cliOptions move in #362, and
resolveBaseURL stays pending #361. CLI > file > default precedence unchanged.
